feat(domain-layer): move Campus.kt to domain/model
