+++ app/src/main/java/com/byansanur/campuslist/data/repository/DataRepository.kt
@@ -3,9 +3,9 @@ package com.byansanur.campuslist.data.repository
 import com.byansanur.campuslist.data.entity.toDomain
 import com.byansanur.campuslist.data.local.CampusLocalRepository
 import com.byansanur.campuslist.data.network.NetworkDataSource
-import com.byansanur.campuslist.domain.Campus
+import com.byansanur.campuslist.domain.model.Campus
 import com.byansanur.campuslist.domain.repository.CampusRepository
-import com.byansanur.campuslist.domain.toLocal
+import com.byansanur.campuslist.domain.model.toLocal
 import com.byansanur.campuslist.utils.flattenToList
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.Dispatchers
