+++ app/src/main/java/com/byansanur/campuslist/domain/model/Campus.kt
@@ -1,4 +1,4 @@
-package com.byansanur.campuslist.domain
+package com.byansanur.campuslist.domain.model
 
 import com.byansanur.campuslist.data.entity.CampusModel
